HCK-14450: Code clean up
