+++ api/fe.js
@@ -1,7 +1,9 @@
 const { generateContainerScript } = require('../forward_engineering/api/generateContainerScript');
 const { isDropInStatements } = require('../forward_engineering/api/isDropInStatements');
+const { generateScript } = require('../forward_engineering/api/generateScript');
 
 module.exports = {
+	generateScript,
 	generateContainerScript,
 	isDropInStatements,
 };
